fix: rename module to diffusionlib
