+++ diffusionlib/conditioning_method.py
@@ -4,7 +4,7 @@ import jax
 import jax.numpy as jnp
 from jaxtyping import Array
 
-from diffusion.measurement_operator import LinearOperator, NoiseOperator
+from diffusionlib.measurement_operator import LinearOperator, NoiseOperator
 
 
 @dataclass
